fixed windows problems

+++ create_mbc.sql
@@ -6,7 +6,7 @@ CREATE TABLE Movies
 (
    id int NOT NULL PRIMARY KEY,
    release_date varchar(10) NOT NULL,
-   revenue int NOT NULL,
+   revenue bigint NOT NULL,
    title varchar(120),
    vote_average float NOT NULL
 );
